Memory management

+++ Kernel/Drivers/keyboard.cpp
@@ -63,7 +63,7 @@ int KeyboardDriver::Read(FILE *file, void *buf, size_t size)
 	char *ptr = (char *)buf;
 
 	// TODO: memcpy
-	while (read < size && pos < buffer_pos)
+	while (read < (int)size && pos < buffer_pos)
 	{
 		ptr[read] = buffer[pos % BUFFER_SIZE];
 		read += 1;
